added something

+++ sources/envelope.c
@@ -13,18 +13,18 @@
 #include "fractol.h"
 
 //sphere - plane - cylinder - bonus: cube
-/*float	envelop_sphere(vec3 pt, t_item item)
+/*double	envelop_sphere(vec3 pt, t_item item)
 {
 	return (vec_dist(pt, item.pos));
 }
 
-float	envelop_plane(vec3 pt, t_item item)
+double	envelop_plane(vec3 pt, t_item item)
 {
 
 }
 
 // cylinder pos is its center, radius is s1, height is s2
-float	envelop_cylinder(vec3 pt, t_item item)
+double	envelop_cylinder(vec3 pt, t_item item)
 {
 	return (vec_dist(pt, item.pos));
 }*/
